+++ common/src/main/java/de/maxhenkel/voicechat/gui/SelectMicrophoneScreen.java
@@ -3,10 +3,7 @@ package de.maxhenkel.voicechat.gui;
 import com.mojang.blaze3d.vertex.PoseStack;
 import de.maxhenkel.voicechat.VoicechatClient;
 import de.maxhenkel.voicechat.gui.widgets.ListScreen;
-import de.maxhenkel.voicechat.voice.client.ALMicrophone;
-import de.maxhenkel.voicechat.voice.client.Client;
-import de.maxhenkel.voicechat.voice.client.ClientManager;
-import de.maxhenkel.voicechat.voice.client.SoundManager;
+import de.maxhenkel.voicechat.voice.client.*;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.screens.Screen;
@@ -44,7 +41,7 @@ public class SelectMicrophoneScreen extends ListScreen<String> {
         Button b = addRenderableWidget(new Button(width / 2 - bw / 2, guiTop + 35, bw, 20, new TranslatableComponent("message.voicechat.select"), button -> {
             VoicechatClient.CLIENT_CONFIG.microphone.set(currentElement).save();
             button.active = false;
-            Client client = ClientManager.getClient();
+            ClientVoicechat client = ClientManager.getClient();
             if (client != null) {
                 client.reloadAudio();
             }