+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/ClientVoicechat.java
@@ -1,0 +1,220 @@
+package de.maxhenkel.voicechat.voice.client;
+
+import de.maxhenkel.voicechat.Voicechat;
+import de.maxhenkel.voicechat.VoicechatClient;
+import de.maxhenkel.voicechat.debug.CooldownTimer;
+import de.maxhenkel.voicechat.intercompatibility.ClientCompatibilityManager;
+import de.maxhenkel.voicechat.intercompatibility.CommonCompatibilityManager;
+import de.maxhenkel.voicechat.voice.common.SoundPacket;
+import net.minecraft.ChatFormatting;
+import net.minecraft.Util;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.network.chat.ComponentUtils;
+import net.minecraft.network.chat.HoverEvent;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.TranslatableComponent;
+
+import javax.annotation.Nullable;
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.UUID;
+
+public class ClientVoicechat {
+
+    private SoundManager soundManager;
+    private final Map<UUID, AudioChannel> audioChannels;
+    private final TalkCache talkCache;
+    @Nullable
+    private MicThread micThread;
+    @Nullable
+    private ClientVoicechatConnection connection;
+    @Nullable
+    private AudioRecorder recorder;
+
+    public ClientVoicechat() {
+        this.talkCache = new TalkCache();
+        try {
+            reloadSoundManager();
+        } catch (SpeakerException e) {
+            e.printStackTrace();
+        }
+        this.audioChannels = new HashMap<>();
+
+        ClientCompatibilityManager.INSTANCE.onVoiceChatConnected(this::startMicThread);
+        ClientCompatibilityManager.INSTANCE.onVoiceChatDisconnected(() -> {
+            stopMicThread();
+            if (connection != null) {
+                connection.close();
+            }
+        });
+    }
+
+    public void connect(InitializationData data) throws IOException {
+        connection = new ClientVoicechatConnection(this, data);
+        connection.start();
+    }
+
+    public void processSoundPacket(SoundPacket packet) {
+        if (connection == null) {
+            return;
+        }
+        synchronized (audioChannels) {
+            if (!ClientManager.getPlayerStateManager().isDisabled()) {
+                AudioChannel sendTo = audioChannels.get(packet.getSender());
+                if (sendTo == null) {
+                    try {
+                        AudioChannel ch = new AudioChannel(this, connection, packet.getSender());
+                        ch.addToQueue(packet);
+                        ch.start();
+                        audioChannels.put(packet.getSender(), ch);
+                    } catch (NativeDependencyException e) {
+                        CooldownTimer.run("decoder_unavailable", () -> {
+                            Voicechat.LOGGER.error("Failed to create audio channel: {}", e.getMessage());
+                            sendPlayerError("messsage.voicechat.playback_unavailable", e);
+                        });
+                    }
+                } else {
+                    sendTo.addToQueue(packet);
+                }
+            }
+
+            audioChannels.values().stream().filter(AudioChannel::canKill).forEach(AudioChannel::closeAndKill);
+            audioChannels.entrySet().removeIf(entry -> entry.getValue().isClosed());
+        }
+    }
+
+    public void reloadSoundManager() throws SpeakerException {
+        if (soundManager != null) {
+            soundManager.close();
+        }
+        soundManager = new SoundManager(VoicechatClient.CLIENT_CONFIG.speaker.get());
+    }
+
+    public void reloadAudio() {
+        Voicechat.LOGGER.info("Reloading audio");
+
+        stopMicThread();
+
+        synchronized (audioChannels) {
+            Voicechat.LOGGER.info("Clearing audio channels");
+            audioChannels.forEach((uuid, audioChannel) -> audioChannel.closeAndKill());
+            audioChannels.clear();
+            try {
+                Voicechat.LOGGER.info("Restarting sound manager");
+                reloadSoundManager();
+            } catch (SpeakerException e) {
+                e.printStackTrace();
+            }
+        }
+
+        Voicechat.LOGGER.info("Starting microphone thread");
+        if (connection != null) {
+            startMicThread(connection);
+        }
+    }
+
+    private void startMicThread(ClientVoicechatConnection connection) {
+        if (micThread != null) {
+            micThread.close();
+        }
+        try {
+            micThread = new MicThread(this, connection);
+            micThread.start();
+        } catch (Exception e) {
+            Voicechat.LOGGER.error("Failed to start microphone thread: {}", e.getMessage());
+            sendPlayerError("messsage.voicechat.microphone_unavailable", e);
+        }
+    }
+
+    public void stopMicThread() {
+        if (micThread != null) {
+            Voicechat.LOGGER.info("Stopping microphone thread");
+            micThread.close();
+            micThread = null;
+        }
+    }
+
+    public void sendPlayerError(String translationKey, Exception e) {
+        LocalPlayer player = Minecraft.getInstance().player;
+        if (player == null) {
+            return;
+        }
+        player.sendMessage(
+                ComponentUtils.wrapInSquareBrackets(new TextComponent(CommonCompatibilityManager.INSTANCE.getModName()))
+                        .withStyle(ChatFormatting.GREEN)
+                        .append(" ")
+                        .append(new TranslatableComponent(translationKey).withStyle(ChatFormatting.RED))
+                        .withStyle(style -> style.withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, new TextComponent(e.getMessage()).withStyle(ChatFormatting.RED))))
+                , Util.NIL_UUID);
+    }
+
+    public void toggleRecording() {
+        setRecording(recorder == null);
+    }
+
+    public void setRecording(boolean recording) {
+        if (recording == (recorder != null)) {
+            return;
+        }
+        LocalPlayer player = Minecraft.getInstance().player;
+        if (recording) {
+            recorder = new AudioRecorder();
+            if (player != null) {
+                player.displayClientMessage(new TranslatableComponent("message.voicechat.recording_started").withStyle(ChatFormatting.DARK_RED), true);
+            }
+        } else {
+            AudioRecorder rec = recorder;
+            recorder = null;
+            if (player != null) {
+                player.displayClientMessage(new TranslatableComponent("message.voicechat.recording_stopped").withStyle(ChatFormatting.DARK_RED), true);
+            }
+            rec.save();
+        }
+    }
+
+    @Nullable
+    public MicThread getMicThread() {
+        return micThread;
+    }
+
+    @Nullable
+    public ClientVoicechatConnection getConnection() {
+        return connection;
+    }
+
+    public SoundManager getSoundManager() {
+        return soundManager;
+    }
+
+    public TalkCache getTalkCache() {
+        return talkCache;
+    }
+
+    @Nullable
+    public AudioRecorder getRecorder() {
+        return recorder;
+    }
+
+    public void close() {
+        synchronized (audioChannels) {
+            Voicechat.LOGGER.info("Clearing audio channels");
+            audioChannels.forEach((uuid, audioChannel) -> audioChannel.closeAndKill());
+            audioChannels.clear();
+        }
+
+        soundManager.close();
+
+        if (micThread != null) {
+            micThread.close();
+        }
+
+        if (recorder != null) {
+            AudioRecorder rec = recorder;
+            recorder = null;
+            rec.close();
+        }
+    }
+
+}