+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/TalkCache.java
@@ -27,7 +27,7 @@ public class TalkCache {
 
     public boolean isTalking(UUID player) {
         if (player.equals(Minecraft.getInstance().player.getUUID())) {
-            Client client = ClientManager.getClient();
+            ClientVoicechat client = ClientManager.getClient();
             if (client != null && client.getMicThread() != null) {
                 return client.getMicThread().isTalking();
             }