+++ fabric/src/main/java/de/maxhenkel/voicechat/events/ClientVoiceChatEvents.java
@@ -1,6 +1,6 @@
 package de.maxhenkel.voicechat.events;
 
-import de.maxhenkel.voicechat.voice.client.Client;
+import de.maxhenkel.voicechat.voice.client.ClientVoicechatConnection;
 import net.fabricmc.fabric.api.event.Event;
 import net.fabricmc.fabric.api.event.EventFactory;
 
@@ -8,8 +8,8 @@ import java.util.function.Consumer;
 
 public class ClientVoiceChatEvents {
 
-    public static final Event<Consumer<Client>> VOICECHAT_CONNECTED = EventFactory.createArrayBacked(Consumer.class, (listeners) -> (client) -> {
-        for (Consumer<Client> listener : listeners) {
+    public static final Event<Consumer<ClientVoicechatConnection>> VOICECHAT_CONNECTED = EventFactory.createArrayBacked(Consumer.class, (listeners) -> (client) -> {
+        for (Consumer<ClientVoicechatConnection> listener : listeners) {
             listener.accept(client);
         }
     });