+++ fabric/src/main/java/de/maxhenkel/voicechat/intercompatibility/FabricClientCompatibilityManager.java
@@ -3,7 +3,7 @@ package de.maxhenkel.voicechat.intercompatibility;
 import com.mojang.blaze3d.platform.InputConstants;
 import de.maxhenkel.voicechat.events.*;
 import de.maxhenkel.voicechat.resourcepacks.IPackRepository;
-import de.maxhenkel.voicechat.voice.client.Client;
+import de.maxhenkel.voicechat.voice.client.ClientVoicechatConnection;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import net.minecraft.client.KeyMapping;
@@ -59,7 +59,7 @@ public class FabricClientCompatibilityManager extends ClientCompatibilityManager
     }
 
     @Override
-    public void emitVoiceChatConnectedEvent(Client client) {
+    public void emitVoiceChatConnectedEvent(ClientVoicechatConnection client) {
         ClientVoiceChatEvents.VOICECHAT_CONNECTED.invoker().accept(client);
     }
 
@@ -69,7 +69,7 @@ public class FabricClientCompatibilityManager extends ClientCompatibilityManager
     }
 
     @Override
-    public void onVoiceChatConnected(Consumer<Client> onVoiceChatConnected) {
+    public void onVoiceChatConnected(Consumer<ClientVoicechatConnection> onVoiceChatConnected) {
         ClientVoiceChatEvents.VOICECHAT_CONNECTED.register(onVoiceChatConnected);
     }